Rewriting the scheduler
This commit rewrites completly the scheduler.

A concept of 'blueprint' has been introduced which holds the graph of
states and edge types. The task holds the blueprint and present a method
_FindPath_ which returns the state reachable from the current state
walking only on a type of edge: mark-based or event-based.

Mutating the task means to find the next state specified by the mark.

The scheduler natural behavior is to push task towards _running_ state
even from _stopped_ or _exited_ state. Only two states are terminal
_inactif_ and _deleted_.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>