Implement evaluation of tasks
This commit implements a naive solution for evaluating tasks.
The idea is that a task is evaluated to true (i.e. set target state to
running) if at least one condition per profile is true.
For example if task has 2 profiles:
- performance:
    - c1
    - c2
- storage:
    - c1
    - c2
and the results of evaluation are:
performance.c1 == true and storate.c2== true => task can be executed

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>