Move retry and cron logic from reconciler to scheduler
This commit moves the logic of retrying and cron from the reconcile to
scheduler. Also, the reconcile function signature changed. It accepts a
single job and return a future. A resolved futures means the reconcile
function return either a _state_ or an _error_.

The reconcile function runs in a go routine but a new reconciliation is
started by the scheduler only when the last future had beed resolved or
there is no future for that job(i.e. the job had never been
reconciled ).

The job struct has been moved to _entity_ package along with its state struct.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>