Add k8S executor
This commit adds a k8s executor.
add:
 - k8s executor
- reconcile function for k8s
The GetEvent of the executor needs recfactoring. At this moment, it just
sends RunningState if the deployment is present. It does not look at
actually state of the deployment.
mod:
 - Kind field for entity.Workload. Configuration mapper set this fields
   from rest response
- signature of GetState and Remove executor changed. It accepts a
  Workload instead of only the id. This is because, the k8s executor
needs the namespace.

Vendor refactoring. The hardware configuration has been removed for now
because the _util_ modules adds dependedies to client-go package
incompatible with the current version of k8s.io/api

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>